requirements.txt
sentence-transformers==2.2.2
umap-learn==0.5.5
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0

matplotlib==3.8.4
streamlit==1.37.0
plotly==5.22.0
